pSwai/requirements-prod.txt
psycopg2-binary
python-ldap
gunicorn
Django
Django-ldap
django_auth_ldap
python-dotenv
python-dotenv[cli]
python-environ
